XML example creation error solved
A mistake within the UML model caused a problem with the xml example creation from the apropiate xsd. In the UML model were some hidden relations between the DGPublication class, DGTransport class and DGAccompanyingInformation class, which caused a circle problem. Therefore the root class could not be identified correctly. Additionally, the aggreation between DGPublication and DGTransport had the wrong direction.